fix signup and remove some relation in entity

+++ src/main/java/com/team/two/mitrais_carrot/service/merchant/BazaarItemService.java
@@ -24,7 +24,7 @@ public class BazaarItemService{
     }
 
     public BazaarItemEntity fetchById(long id){
-        return bazaarItemRepository.findById(id).orElse(null);
+        return bazaarItemRepository.findById((int) id).orElse(null);
     }
